Refactoring and fix tests.
